nostrdb: port everything over to be in as sync as possible
for now

+++ nostrdb/threadpool.h
@@ -87,7 +87,7 @@ static inline void threadpool_destroy(struct threadpool *tp)
 {
 	struct thread *t;
 
-	for (uint64_t i = 0; i < tp->num_threads; i++) {
+	for (int i = 0; i < tp->num_threads; i++) {
 		t = &tp->pool[i];
 		if (!prot_queue_push(&t->inbox, tp->quit_msg)) {
 			pthread_exit(&t->thread_id);
